separated Root for easier fast refresh, saving csrf cookie client side for debugging
